remove redundant utils
